requirements.txt
python-dotenv==0.14.0
kafka-python==2.0.2
